Rename .jscad to .js to support proper mime type inference.
I'm working to update the renderer link the README to work in current
browser versions. Since the `importScripts` API explicitly requires
specific mime types to execute scripts. The Github raw content serving
instructs the browser not to infer the mime type from the content, and
since .jscad isn't an official extension for `application/javascript`,
the code isn't executing. So, rename the script extensions to .js since
they are, in fact, JavaScript.

For some reason with this change the exports shim I'd hacked together
stopped working for test runs in csg_dependency_graph.js[cad]. I'm not
sure if something starting doing additional transpiling, or if something
else caused it. I did confirm that the previous revision did run the
tests successfully so it was definitely only the file extension change
breaking things. That being said, I realized that the shim had a bug
anyway: JS's variable hoisting meant the shimmed `exports` var was being
setup regardless of whether the `if` condition was satisfied. Changing
things to use an intermediate var either way resolved the issue.