Use jsdlivr CDN to serve code to renderer.
The Github raw content serving issues a content type header of
`text/plain` and also instructs the browser not to sniff the actual mime
type. That breaks the `importScripts` API, so the 3D model viewer
doesn't work.

The [jsDelivr](https://www.jsdelivr.com/) CDN allows you to reference
Github sources, and it properly sets the mime type headers based on file
extension. So, use that as an intermediary to get rendering working
again.