Redraw switch matrix/switch matrix components boundary.
It turns out that many of methods/properties here are really about
drawing specific components (not necessarily matrix creation and
layout), and we already have this container class for those components
anyway, so instead of treating it mostly like a bag of properties, we
can actually move responsibility for building those components into the
SwitchMatrixComponents class.

It'ss also a clear win for separation of concerns since several of the
arguments passed to the SwitchMatrix class were solely about the
components, and just got wholesale passed on through.

This should make it easier to reuse the SwitchMatrix class for the
combined keyboard, since it appears likely that it will need to use a
fairly different strategy for building its components.