Layout keyboard objects relative to switches.
Previously we layed out the switches, generated the various plate
components (for each of the two matrices independently), and then
layed out those plate components relative to each other. This approach
worked fine for the current keyboard style so far, but in preparing to
add a combined board style, it's become obvious that it's cleaner (and
more extensible) to lay out the thumb matrix switches relative to the
primary matrix switches first. Then the components will be layed out
properly by extension, and we get the benefit of being able to work with
the switch hole objects directly (both to manually transform them and to
consume properties) knowing that they're all already in the proper
positions.

Now that we're connecting the plate to the switches, we had to
additionally tweak the connector code on several of the plate
components. Those connectors were previously just assuming that they
only needed to be correct relative to other plate components, but now
they need to actually specify points relative to the key switch holes
they're technically based off of. For example, the key switch holes are
generated as a radius around Z=0, but the plate components are generated
as extrusions starting at Z=0, so the points in absolute terms at
initial generation are going to be offset by one half of the switch
thickness (i.e., the aforementioned radius).

This change resulted in extremely minor changes to the magic numbers in
the sanity spec, but I've visually confirmed that the resulting model
looks correct, and the changes are mathematically speaking not
meaningful.